Move pipeline-related config back to main pipeline.py file.
